An S3 class constructor of representing Venn plot components.

An S3 class constructor of representing Venn plot components.

VennPlotData(x)

Arguments

x

data source of a VennPlotData object

Slots

shapeId

shape id

type

type of shape

nsets

number of sets

setEdge

a data.frame, the coordinates of set edges, can be retrieved by venn_setedge()

setLabel

a data.frame, the coordinates of set labels, can be retrieved by venn_setlabel()

regionEdge

a data.frame, the coordinates of different regions, can be retrieved by venn_regionedge()

regionLabel

a data.frame, the centroid of the regions, where region labels anchored, can be retrieved by venn_regionlabel()

setData

a data.frame, the set data provided by user, can be retrieved by venn_set()

regionData

a data.frame, the region data that calculated by ggVennDiagram, can be retrieved by venn_region()