get plot data

get plot data

process_data(venn, nsets = NULL, shape_id = NULL, type = NULL)

# S4 method for Venn
process_data(venn, nsets = length(venn@sets), shape_id = NULL, type = NULL)

Arguments

venn

a Venn object

nsets

This parameter will be set automatically.

shape_id

apply filter to internal shapes. i.e. shape_id = "601"

type

apply filter to internal shapes. i.e. type = "polygon"

Details

This function will conduct set operations and combine the outputs will stored shapes, thus produce a dataset for plot in next step.

Run `get_shapes()` to show all the characteristics of available shapes. Run `plot_shapes()` to view those shapes.

Examples

if (FALSE) {
 venn = Venn(list(A=1:3,B=2:5,C=4:8))
 data = process_data(venn)
}